Use python executable from virtualenv
This removes the need for all the activation code, greatly simplifying
some stuff. It is though a bit harder to test, so some unit tests was
removed in favor of a more integration tests. Running the integration
tests on a given platform can be done by performing the same steps done by
travis.

Since many of the unit tests relied on the test_venv to be ignored, added
another test module for testing exclusion filters.